- Create new folder under gillespy2/solvers, named utilities - Move Tau file to utilities folder - create solverutils.py, which contains: functions used in the C solvers, and well as functions used in the numpy solvers
