+++ gillespy2/solvers/utilities/Tau.py
